main function of the project
main.c contains main function of the project which covers all possible programs running on the microwave. 
programs.h file contains the necessary header files and the prototypes of the needed fucntions in main to make the program more modular. 
programs.c file contains the implementation of the functions in programs.h